+++ backend/src/main/java/com/vehicle/marketplace/model/request/CarUpdateRequest.java
@@ -14,7 +14,7 @@ public class CarUpdateRequest {
     String model;
     Integer manufactureYear;
     String licensePlate;
-    Integer price;
+    Long price;
     Integer count;
     String status;
     String color;
