+++ backend/src/main/java/com/vehicle/marketplace/model/response/CarResponse.java
@@ -14,7 +14,7 @@ public class CarResponse {
     String model;
     Integer manufactureYear;
     String licensePlate;
-    Integer price;
+    Long price;
     Integer count;
     String status;
     String color;
